refactor: extract project session from initMutation
openProjectSession owns per-project loading and the project-scoped
subscriptions (audio sync, debounced auto-save), and returns a
symmetric dispose() (unsubscribe, flush timer, clear undo history).
Nothing calls dispose() yet; it is the seam for switching projects
without a page reload (#152).
